dog info

+++ 0x0E-structures_typedef/dog.h
@@ -1,21 +1,21 @@
 #ifndef DOG_H
 #define DOG_H
-/**
- * struct dog - struct dog info
- *
- * @name: dog name
- * @age: dog age
- * @owner: dog owner
- *
- * Description: struct of dog
- */
 
+/**
+ * struct dog - a dog's basic info
+ * @name: First member
+ * @age: Second member
+ * @owner: Third member
+ *
+ * Description: Longer description
+ */
 struct dog
 {
 	char *name;
-	int age;
+	float age;
 	char *owner;
 };
+
 /**
  * dog_t - typedef for struct dog
  */
@@ -29,4 +29,5 @@ char *_strcpy(char *dest, char *src);
 int _strlen(char *s);
 
 #endif
+
